Add CI test for uniqueness of mock-server testcases.
