+++ mock-recipe-server/pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+addopts = -p no:django
